started usbstreamer for linux environment
